tests: Port kubelet tests to Windows
Ports kubelet/util unit tests to Windows.
